requirements.txt
pip==9.0.1
Flask==0.12.2
slackclient==1.0.6
requests==2.18.1
python-pidfile==1.0.1

